REST parameter map can now support any type of RPC parameter
Especially it is no longer bound to a be a string. Actually this is a
breaking change but currently no one is using it: any parameter must be
JSON serialized, i.e. a number 1 becomes the string "1" and especially
a string "abc" becomes the string "\"abc\"". The decoding algoritm can
handle JSON array and object as well - the later one will become
QVariantMap after deserialization.